add balance view and move category list to own view.
