+++ Src/MoneyFox/UserControls/BalanceUserControl.xaml.cs
@@ -1,0 +1,13 @@
+﻿using Xamarin.Forms.Xaml;
+
+namespace MoneyFox.UserControls
+{
+	[XamlCompilation(XamlCompilationOptions.Compile)]
+	public partial class BalanceUserControl
+	{
+		public BalanceUserControl ()
+		{
+			InitializeComponent ();
+		}
+	}
+}
