+++ Src/MoneyFox/UserControls/CategoryListUserControl.xaml
@@ -1,0 +1,89 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<ContentView xmlns="http://xamarin.com/schemas/2014/forms" 
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:suave="clr-namespace:SuaveControls.Views;assembly=SuaveControls.FloatingActionButton"
+             xmlns:behavior="clr-namespace:MoneyFox.Behavior;assembly=MoneyFox"
+             x:Class="MoneyFox.UserControls.CategoryListUserControl">
+
+  <ContentView.Content>
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="60" />
+                <RowDefinition Height="240*" />
+            </Grid.RowDefinitions>
+
+            <SearchBar x:Name="SearchBar" 
+                       Grid.Row="0"
+                       VerticalOptions="Start" 
+                       Placeholder="{Binding Resources[SearchLabel]}"
+                       SearchCommand="{Binding SearchCommand}"
+                       SearchCommandParameter="{Binding Text, Source={x:Reference SearchBar}}" >
+                <SearchBar.Behaviors>
+                    <behavior:TextChangedBehavior />
+                </SearchBar.Behaviors>
+            </SearchBar>
+
+            <ListView x:Name="CategoryList"
+                      Grid.Row="1"
+                      ItemsSource="{Binding CategoryList}"
+                      HasUnevenRows="True"
+                      CachingStrategy="RecycleElement"
+                      IsGroupingEnabled="True"
+                      GroupShortNameBinding="{Binding Key}"
+                      GroupDisplayBinding="{Binding Key}"
+                      SeparatorVisibility="None">
+                <ListView.GroupHeaderTemplate>
+                    <DataTemplate>
+                        <ViewCell>
+                            <StackLayout>
+                                <Label Text="{Binding Key}" Style="{StaticResource GroupHeaderStyle}" />
+                                <BoxView HeightRequest="1" HorizontalOptions="FillAndExpand" Color="LightGray" />
+                            </StackLayout>
+                        </ViewCell>
+                    </DataTemplate>
+                </ListView.GroupHeaderTemplate>
+                <ListView.ItemTemplate>
+                    <DataTemplate>
+                        <ViewCell>
+                            <ViewCell.ContextActions>
+                                <MenuItem Clicked="EditCategory" CommandParameter="{Binding .}"
+                                          Text="More" />
+                                <MenuItem Clicked="DeleteCategory" CommandParameter="{Binding .}"
+                                          Text="Delete" IsDestructive="True" />
+                            </ViewCell.ContextActions>
+                            <StackLayout Margin="{StaticResource NegativeMediumLeftRightMargin}" Spacing="0">
+                                <StackLayout BackgroundColor="{StaticResource ListItemColor}"
+                                             Orientation="Vertical"
+                                             Padding="3"
+                                             Spacing="{StaticResource DefaultSpacing}"
+                                             Margin="{StaticResource DefaultListItemsMargin}">
+                                    <Label Text="{Binding Name}" Style="{StaticResource ListItemHeaderStyle}" />
+                                </StackLayout>
+                            </StackLayout>
+                        </ViewCell>
+                    </DataTemplate>
+                </ListView.ItemTemplate>
+            </ListView>
+
+            <suave:FloatingActionButton Grid.Row="1" 
+                                           x:Name="FAB" 
+                                           HorizontalOptions="EndAndExpand" 
+                                           WidthRequest="50" 
+                                           HeightRequest="50"
+                                           Margin="12"
+                                           VerticalOptions="EndAndExpand"
+                                           Image="ic_add.png" 
+                                           ButtonColor="{StaticResource AccentColor}" 
+                                           Clicked="AddCategoryClick"/>
+
+            <Label Grid.Row="1"
+                   Margin="6"
+                   HorizontalOptions="Center"
+                   VerticalOptions="Center"
+                   Style="{StaticResource DeemphasizedLabelStyle}"
+                   LineBreakMode="WordWrap"
+                   IsVisible="{Binding IsCategoriesEmpty}"
+                   Text="{Binding Resources[NoCategoriesMessage]}"/>
+        </Grid>
+    </ContentView.Content>
+</ContentView>
