+++ Src/MoneyFox/Views/CategoryListPage.xaml
@@ -2,101 +2,15 @@
 <d:MvxContentPage x:TypeArguments="viewModels:CategoryListViewModel" xmlns="http://xamarin.com/schemas/2014/forms"
                   xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
                   xmlns:d="clr-namespace:MvvmCross.Forms.Views;assembly=MvvmCross.Forms"
-                  xmlns:behavior="clr-namespace:MoneyFox.Behavior;assembly=MoneyFox"
                   xmlns:moneyFox="clr-namespace:MoneyFox;assembly=MoneyFox"
-                  xmlns:controls="clr-namespace:SuaveControls.Views;assembly=SuaveControls.FloatingActionButton"
+                  xmlns:views="clr-namespace:MoneyFox.UserControls;assembly=MoneyFox"
                   xmlns:viewModels="clr-namespace:MoneyFox.Business.ViewModels;assembly=MoneyFox.Business"
                   x:Class="MoneyFox.Views.CategoryListPage"
                   DataContext="{x:Static moneyFox:DesignTimeViewModelLocator.DesignTimeCategoryListViewModel}">
 
     <ContentPage.Content>
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="60" />
-                <RowDefinition Height="240*" />
-            </Grid.RowDefinitions>
-
-            <StackLayout Grid.Row="0">
-                <StackLayout.Margin>
-                    <OnPlatform x:TypeArguments="Thickness">
-                        <On Platform="iOS" Value="0" />
-                        <On Platform="UWP" Value="{StaticResource MediumLeftRightMargin}" />
-                        <On Platform="Android" Value="0" />
-                    </OnPlatform>
-                </StackLayout.Margin>
-
-                <Label Style="{StaticResource DefaultTextStyle}">Search</Label>
-
-                <SearchBar x:Name="SearchBar" VerticalOptions="Start" 
-                           SearchCommand="{Binding SearchCommand}"
-                           SearchCommandParameter="{Binding Text, Source={x:Reference SearchBar}}" >
-                    <SearchBar.Behaviors>
-                        <behavior:TextChangedBehavior />
-                    </SearchBar.Behaviors>
-                </SearchBar>
-            </StackLayout>
-
-            <ListView x:Name="CategoryList"
-                      Grid.Row="1"
-                      ItemsSource="{Binding CategoryList}"
-                      HasUnevenRows="True"
-                      CachingStrategy="RecycleElement"
-                      IsGroupingEnabled="True"
-                      GroupShortNameBinding="{Binding Key}"
-                      GroupDisplayBinding="{Binding Key}"
-                      SeparatorVisibility="None">
-                <ListView.GroupHeaderTemplate>
-                    <DataTemplate>
-                        <ViewCell>
-                            <StackLayout>
-                                <Label Text="{Binding Key}" Style="{StaticResource GroupHeaderStyle}" />
-                                <BoxView HeightRequest="1" HorizontalOptions="FillAndExpand" Color="LightGray" />
-                            </StackLayout>
-                        </ViewCell>
-                    </DataTemplate>
-                </ListView.GroupHeaderTemplate>
-                <ListView.ItemTemplate>
-                    <DataTemplate>
-                        <ViewCell>
-                            <ViewCell.ContextActions>
-                                <MenuItem Clicked="EditCategory" CommandParameter="{Binding .}"
-                                          Text="More" />
-                                <MenuItem Clicked="DeleteCategory" CommandParameter="{Binding .}"
-                                          Text="Delete" IsDestructive="True" />
-                            </ViewCell.ContextActions>
-                            <StackLayout Margin="{StaticResource NegativeMediumLeftRightMargin}" Spacing="0">
-                                <StackLayout BackgroundColor="{StaticResource ListItemColor}"
-                                             Orientation="Vertical"
-                                             Padding="3"
-                                             Spacing="{StaticResource DefaultSpacing}"
-                                             Margin="{StaticResource DefaultListItemsMargin}">
-                                    <Label Text="{Binding Name}" Style="{StaticResource ListItemHeaderStyle}" />
-                                </StackLayout>
-                            </StackLayout>
-                        </ViewCell>
-                    </DataTemplate>
-                </ListView.ItemTemplate>
-            </ListView>
-
-            <controls:FloatingActionButton Grid.Row="1" 
-                                           x:Name="FAB" 
-                                           HorizontalOptions="EndAndExpand" 
-                                           WidthRequest="50" 
-                                           HeightRequest="50"
-                                           Margin="12"
-                                           VerticalOptions="EndAndExpand"
-                                           Image="ic_add.png" 
-                                           ButtonColor="{StaticResource AccentColor}" 
-                                           Clicked="AddCategoryClick"/>
-
-            <Label Grid.Row="1"
-                   Margin="6"
-                   HorizontalOptions="Center"
-                   VerticalOptions="Center"
-                   Style="{StaticResource DeemphasizedLabelStyle}"
-                   LineBreakMode="WordWrap"
-                   IsVisible="{Binding IsCategoriesEmpty}"
-                   Text="{Binding Resources[NoCategoriesMessage]}"/>
-        </Grid>
+        <StackLayout>
+            <views:CategoryListUserControl />
+        </StackLayout>
     </ContentPage.Content>
 </d:MvxContentPage>
